+++ config/hard/ErdősNumber/code/CGCode.js
@@ -1,0 +1,13 @@
+const scientist = readline();
+const N = parseInt(readline());
+for (let i = 0; i < N; i++) {
+    const title = readline();
+}
+for (let i = 0; i < N; i++) {
+    const authors = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');