+++ config/hard/ErdősNumber/code/CGCode.ts
@@ -1,0 +1,13 @@
+const scientist: string = readline();
+const N: number = parseInt(readline());
+for (let i = 0; i < N; i++) {
+    const title: string = readline();
+}
+for (let i = 0; i < N; i++) {
+    const authors: string = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');